채팅 리스트 로직 수정

+++ src/main/java/com/example/DummyTalk/User/Repository/UserRepository.java
@@ -12,5 +12,5 @@ public interface UserRepository extends JpaRepository<User, Long> {
 
     boolean existsByUserEmail(String email);
 
-    User findByUserId(int sender);
+    User findByUserId(Long sender);
 }
